feat(teams):supprimer des équipes qui n'ont pas de matchs affectés

+++ PhP/lamp/www/view/teams.php
@@ -8,9 +8,26 @@ ob_start();
 <ul>
 
 <?php
+
+$games = VolscoreDB::getGames();
+
 foreach ($teams as $team)
 {
-    echo "<li>".$team->name."<a href='?action=deleteTeam&teamid=".$team->id."' class='btn')'><img src='../images/delete.png' width='20'></a> </li>";
+    $isDeletable = true;
+    //création du bouton si l'équipe n'a aucun match planifié
+    foreach($games as $game){
+        if(!($game->receivingTeamId == $team->id) && !($game->visitingTeamId == $team->id)){
+            $isDeletable = true;
+        } else {
+            $isDeletable = false;
+            break;
+        }
+    }
+    if($isDeletable){
+        echo "<li>".$team->name."<a href='?action=deleteTeam&teamid=".$team->id."' class='btn')'><img src='../images/delete.png' width='20'></a> </li>";
+    } else {
+        echo "<li>".$team->name."</li>";
+    }
 }
 ?>
 </ul>
